==> r/1_npdr_cc.Rmd <==
---
title: "Comparing NPDR, STIR and Relief-F"
output: html_document
---

### Load packages:

```{r echo=FALSE}
library(privateEC)
library(broom)
library(tidyverse)
# devtools::install_github("insilico/npdr") # npdr install
library(npdr)

rm(list = ls())
set.seed(1618)

cbbPalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", 
                "#0072B2", "#D55E00", "#CC79A7", "#c5679b", "#be548f")
```


### Simulate case-control interaction effect data:

```{r}
#####
n.samples <- 300     # 100 samples in train/holdout
n.variables <- 1000   # 100 features
label <- "class"     # tells simulator to do case/control and adds this colname
type <- "interactionErdos" # or mainEffect
# type <-"mainEffect"
bias <- 0.4          # moderate effect size
pct.signals <- 0.1   # pct functional features
alpha <- 0.05
verbose <- FALSE
cc.3sets <- createSimulation(num.samples = n.samples,
                              num.variables = n.variables,
                              pct.signals = pct.signals,
                              label = label,
                              bias = bias,
                              pct.train = 1/3,
                              pct.holdout = 1/3,
                              pct.validation = 1/3,
                              sim.type = type,
                              save.file = NULL,
                              verbose = verbose)
# combine train and holdout into 200 samples x 100 attributes
# ignore validation set for now; cc = case.control (binary outcome)
cc.data <- rbind(cc.3sets$train, cc.3sets$holdout)
n.samples.cc <- nrow(cc.data)
pheno.cc <- as.factor(cc.data[,label])
functional.cc <- cc.3sets$signal.names # functional attributes
```


### Run npdr:

```{r}
#####
system.time(
  npdr.cc.results <- 
  npdr(label, cc.data, regression.type="glm", 
          attr.diff.type = "numeric-abs", nbd.method = "multisurf", 
          nbd.metric = "manhattan", msurf.sd.frac = .5, 
          padj.method = "bonferroni", verbose = T) %>%
  # rownames_to_column('att') %>%
  mutate(npdr.log10 = -log10(pval.att))
)

# get functional attribute detection stats
npdr.cc.positives <- npdr.cc.results %>%
  filter(pval.adj < alpha) %>%
  pull(att)
npdr.cc.detect.stats <- detectionStats(functional.cc, npdr.cc.positives)
cat(npdr.cc.detect.stats$report)

```

### Univariate regression:

```{r}
system.time(
  gene_mdd <- uniReg(
  outcome = label, 
  dataset = cc.data, 
  regression.type = 'binomial') %>%
  data.frame() %>%
  rownames_to_column('att') %>%
  mutate(functional = grepl('sim', att))
)
```

### Comparing univariate regression and NPDR:

```{r}
n.pos <- gene_mdd %>%
  filter(p.adj < alpha) %>%
  nrow()
true.pos <- gene_mdd %>%
  filter((p.adj < alpha) & functional) %>%
  nrow()
(uni.recall <- true.pos/(n.variables*pct.signals))
(uni.precis <- true.pos/n.pos)

test.df <- npdr.cc.results %>%
  mutate(functional = att %in% functional.cc) 
true.pos <- test.df %>%
  filter((pval.adj < alpha) & functional) %>%
  nrow()
n.pos <- test.df %>%
  filter((pval.adj < alpha)) %>%
  nrow()
(npdr.recall <- true.pos/(n.variables*pct.signals))
(npdr.precis <- true.pos/n.pos)

pr_uni <- PRROC::pr.curve(scores.class0 = gene_mdd %>%
                        filter(functional == T) %>%
                        pull(beta.Z.att) %>% abs(),
                      scores.class1 = gene_mdd %>%
                        filter(functional == F) %>%
                        pull(beta.Z.att) %>% abs(),
                      curve = T)

pr_npdr <- PRROC::pr.curve(scores.class0 = test.df %>%
                        filter(functional == T) %>%
                        pull(beta.Z.att),
                      scores.class1 = test.df %>%
                        filter(functional == F) %>%
                        pull(beta.Z.att),
                      curve = T)
pr_uni$auc.integral
pr_npdr$auc.integral

```


### Run STIR:

```{r}
# expect: npdr gives same resutls as original t-STIR
# stir interface requires splitting phenotype and predictor matrix, 
# and requires finding the neighborhood separately.
predictors.cc.mat <- cc.data %>% dplyr::select(-class)
neighbor.idx.observed <- stir::find.neighbors(predictors.cc.mat, 
                                        pheno.cc, k = 0, method = "multisurf")
stir.results <- stir::stir(predictors.cc.mat, neighbor.idx.observed, k = 0, 
                     metric = "manhattan", method = "multisurf")
multisurf_softed_imp <- stir.results$STIR_T %>% 
  rownames_to_column('att') %>%
  mutate(stir.log10 = -log10(t.pval))

# functional attribute detection stats
tstat_stir.detect.stats <- detectionStats(
  functional.cc, multisurf_softed_imp %>%
    filter(t.pval.adj < alpha) %>%
    pull(att))
cat(tstat_stir.detect.stats$report)
```

### Compare NPDR and STIR
```{r}
pcutoff <- -log10(alpha/n.variables) # bonferoni adjustment
test.df <- merge(multisurf_softed_imp, npdr.cc.results) %>% 
  mutate(functional = att %in% functional.cc)
r2_stir_cc <- round(cor(test.df$stir.log10, test.df$npdr.log10), digits = 3)

npdr_stir <- test.df
```


### Compute Relief-F score with the `CORElearn` package 
(fixed $k$ with theoretical surf value)

```{r}
core.learn.cc <- CORElearn::attrEval(label, data = cc.data,
                                      estimator = "ReliefFequalK",
                                      costMatrix = NULL,
                                      outputNumericSplits=FALSE,
                                      kNearestEqual = knnSURF(n.samples.cc,.5))
# core.learn.cc.order <- order(core.learn.cc, decreasing = T)
corelearn.df <- data.frame(att=names(core.learn.cc), rrelief=core.learn.cc)
last.imp.att <- which(npdr.cc.results$pval.adj>alpha)[1]-1
npdr.beta.cutoff.cc <- npdr.cc.results[last.imp.att, 'beta.Z.att']
```

For later ploting:

```{r}
test.df <- merge(corelearn.df, npdr.cc.results) %>% 
  mutate(functional = att %in% functional.cc)
r2_cc <- round(cor(test.df$rrelief, test.df$beta.Z.att), digits = 3)
```


Draw the Precision-Recall curve:

```{r}
pr_npdr <- PRROC::pr.curve(scores.class0 = test.df %>% 
                        filter(functional == T) %>% 
                        pull(beta.Z.att),
                      scores.class1 = test.df %>% 
                        filter(functional == F) %>% 
                        pull(beta.Z.att), 
                      curve = T)

pr_relief <- PRROC::pr.curve(scores.class0 = test.df %>% 
                        filter(functional == T) %>% 
                        pull(rrelief), 
                      scores.class1 = test.df %>% 
                        filter(functional == F) %>% 
                        pull(rrelief), 
                      curve = T)

x <- as.data.frame(pr_npdr$curve) 
colnames(x) <- paste0(c('Recall', 'Precision', 'w'))
y <- as.data.frame(pr_relief$curve) 
colnames(y) <- paste0(c('Recall', 'Precision', 'w'))
auprc_npdr_cc <- round(pr_npdr$auc.integral, 3)
auprc_relief_cc <- round(pr_relief$auc.integral, 3)

prr.dat <- rbind(x, y) %>% 
  mutate(type = c(rep('NPDR', nrow(x)), rep('Relief-F', nrow(y))))
```



How much better is NPDR? Is this just by chance? 100 replications result show the same thing (npdr_100_cc.Rmd).




Compare with random forest:

```{r}
cc_forest <- randomForest::randomForest(class ~ ., data = cc.data, importance = T)
rf_imp_cc <- randomForest::importance(cc_forest) %>%
  data.frame() %>%
  rownames_to_column('att') %>%
  mutate(functional = att %in% functional.cc)
rf_npdr_cc <- merge(rf_imp_cc, npdr.cc.results, by = 'att') %>%
  mutate(functional = att %in% functional.cc)
r2_rf_cc <- round(cor(rf_npdr_cc$MeanDecreaseAccuracy, rf_npdr_cc$beta.Z.att), 3)
```

Draw the Precision-Recall curve:

```{r}
pr_rf_cc <- PRROC::pr.curve(scores.class0 = rf_imp_cc %>% 
                        filter(functional == T) %>% 
                        pull(MeanDecreaseAccuracy), 
                      scores.class1 = rf_imp_cc %>% 
                        filter(functional == F) %>% 
                        pull(MeanDecreaseAccuracy), 
                      curve = T)

z <- as.data.frame(pr_rf_cc$curve) 
colnames(z) <- paste0(c('Recall', 'Precision', 'w'))
auprc_rf_cc <- round(pr_rf_cc$auc.integral, 3)

prr_rf_cc <- rbind(x, z) %>% 
  mutate(type = c(rep('NPDR', nrow(x)), rep('Random forest', nrow(z))))

save(rf_imp_cc, auprc_rf_cc, pr_rf_cc, prr_rf_cc, rf_npdr_cc, r2_rf_cc,
     file = 'results/rf_cc.Rdata')
```


What about auROC?

```{r}
ss_npdr <- pROC::roc(test.df$functional, test.df$beta.Z.att)
ss_relief <- pROC::roc(test.df$functional, test.df$rrelief)
ss_rf <- pROC::roc(rf_imp_cc$functional, rf_imp_cc$MeanDecreaseAccuracy)
auroc_npdr_cc <- ss_npdr$auc
auroc_relief_cc <- ss_relief$auc
auroc_rf_cc <- ss_rf$auc

ss_df_cc <- data.frame(
  Sensitivity = c(
    ss_npdr$sensitivities,
    ss_relief$sensitivities,
    ss_rf$sensitivities
  ),
  Specificity = c(
    ss_npdr$specificities,
    ss_relief$specificities,
    ss_rf$specificities
  )
) %>%
  mutate(type = c(
    rep('NPDR', length(ss_npdr$sensitivities)),
    rep('Relief', length(ss_relief$sensitivities)),
    rep('Random forest', length(ss_rf$sensitivities))
  ),
  pct_selected = c(
    seq(1, 0, length.out = length(ss_npdr$sensitivities)),
    seq(1, 0, length.out = length(ss_relief$sensitivities)),
    seq(1, 0, length.out = length(ss_rf$sensitivities))
  ))
```


```{r}
test_df_cc <- test.df
prr_dat_cc <- prr.dat

save(npdr_stir, ss_df_cc, test_df_cc, prr_dat_cc,
     auprc_npdr_cc, auprc_relief_cc, r2_cc,
     npdr.beta.cutoff.cc, ss_df_cc,
     auroc_npdr_cc, auroc_relief_cc,
     auroc_rf_cc, pcutoff,
     file = 'results/df_cc.Rdata')
```

```{r}
sessionInfo()
```
