==> r/2_npdr_relief_100_qtrait.Rmd <==
---
title: "Comparing NPDR and Relief-F, 100 replications"
output: html_document
---

### Load packages:

```{r echo=FALSE}
library(privateEC)
library(broom)
library(tidyverse)
# library(devtools)
# install_github("insilico/npdr") # npdr install
library(npdr)

rm(list = ls())
set.seed(1618)

cbbPalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", 
                "#0072B2", "#D55E00", "#CC79A7", "#c5679b", "#be548f")
```


### Simulate case-control interaction effect data:

```{r}
#####

##### simulate case-control interaction effect data 
n.samples <- 300     # 100 samples in train/holdout/test
n.variables <- 1000   # 100 features
label <- "qtrait"   # tells simulator to do quantitative trait and adds this colname
type <- "mainEffect"
bias <- 0.6          # moderate effect size
pct.signals <- 0.1   # pct functional features
verbose <- FALSE
```


```{r}
n_sims <- 100
r2_vec <- vector(mode = 'numeric', length = n_sims)
r2_rf_vec <- vector(mode = 'numeric', length = n_sims)
pr_npdr_vec <- vector(mode = 'numeric', length = n_sims)
pr_relief_vec <- vector(mode = 'numeric', length = n_sims)
pr_rf_vec <- vector(mode = 'numeric', length = n_sims)

set.seed(16183)
for (i in 1:n_sims){
  qtrait.3sets <- createSimulation(num.samples = n.samples,
                              num.variables = n.variables,
                              pct.signals = pct.signals,
                              label = label,
                              bias = bias,
                              pct.train = 1/3,
                              pct.holdout = 1/3,
                              pct.validation = 1/3,
                              sim.type = type,
                              save.file = NULL,
                              verbose = verbose)
  qtrait.data <- rbind(qtrait.3sets$train, qtrait.3sets$holdout)
  n.samples.qtrait <- nrow(qtrait.data)
  functional.qtrait <- qtrait.3sets$signal.names # functional attributes

  npdr.qtrait.results <- 
  npdr(label, qtrait.data, regression.type="lm", 
          attr.diff.type = "numeric-abs", nbd.method = "multisurf", 
          nbd.metric = "manhattan", msurf.sd.frac = .5, 
          padj.method = "bonferroni", verbose = T) %>%
  mutate(npdr.log10 = -log10(pval.att))
  
  core.learn.qtrait <- CORElearn::attrEval(label, data = qtrait.data,
                                      estimator = "RReliefFequalK",
                                      costMatrix = NULL,
                                      outputNumericSplits=FALSE,
                                      kNearestEqual = knnSURF(n.samples.qtrait,.5))
  corelearn.df <- data.frame(att=names(core.learn.qtrait), rrelief=core.learn.qtrait)
  
  test.df <- merge(corelearn.df, npdr.qtrait.results) %>% 
    mutate(functional = att %in% functional.qtrait)
  qtrait_forest <- randomForest::randomForest(qtrait ~ ., data = qtrait.data, importance = T)
  rf_imp_qtrait <- randomForest::importance(qtrait_forest) %>%
    data.frame() %>%
    rownames_to_column('att') %>%
    mutate(functional = att %in% functional.qtrait)
  rf_npdr_qtrait <- merge(rf_imp_qtrait, npdr.qtrait.results, by = 'att') 
  
  r2_rf_vec[i] <- round(cor(rf_npdr_qtrait$X.IncMSE, rf_npdr_qtrait$beta.Z.att), 3)

  r2_vec[i] <- round(cor(test.df$rrelief, test.df$beta.Z.att), digits = 3)
  
  # compute area under the precision-recall curve
  pr_npdr <- PRROC::pr.curve(scores.class0 = test.df %>% 
                        filter(functional == T) %>% 
                        pull(beta.Z.att),
                      scores.class1 = test.df %>% 
                        filter(functional == F) %>% 
                        pull(beta.Z.att), 
                      curve = T)
  pr_relief <- PRROC::pr.curve(scores.class0 = test.df %>% 
                        filter(functional == T) %>% 
                        pull(rrelief), 
                      scores.class1 = test.df %>% 
                        filter(functional == F) %>% 
                        pull(rrelief), 
                      curve = T)
  pr_rf <- PRROC::pr.curve(scores.class0 = rf_imp_qtrait %>% 
                        filter(functional == T) %>% 
                        pull(X.IncMSE), 
                      scores.class1 = rf_imp_qtrait %>% 
                        filter(functional == F) %>% 
                        pull(X.IncMSE), 
                      curve = T)
  
  pr_npdr_vec[i] <- pr_npdr$auc.integral
  pr_relief_vec[i] <- pr_relief$auc.integral
  pr_rf_vec[i] <- pr_rf$auc.integral

}

```


```{r}
pr_df <- data.frame(aupr = c(pr_relief_vec, pr_npdr_vec, pr_rf_vec),
                    type = rep(c('RRelief', 'NPDR', 'Random forest'), each = n_sims))
pr_df$type <- factor(pr_df$type, levels = rev(levels(pr_df$type)))
summary(r2_vec)
summary(r2_rf_vec)
pr_compare_relief <- t.test(pr_npdr_vec, pr_relief_vec)
pr_compare_rf <- t.test(pr_npdr_vec, pr_rf_vec)
print(pr_compare_relief$p.value)
pr_p <- round(pr_compare_relief$p.value, 3)
# pr_compare_qtrait <- pr
print(pr_compare_rf$p.value)
pr_p <- round(pr_compare_rf$p.value, 3)
save.image(file = 'results/npdr_100_qtrait.Rdata')

```
