==> r/3_visualize_simulation.Rmd <==
---
title: "Plots together"
output: html_document
---

Combine plots together:

```{r}
# rm(list = ls())
load('results/df_cc.Rdata')
load('results/df_qtrait.Rdata')
load('results/rf_cc.Rdata')
load('results/rf_qtrait.Rdata')

library(tidyverse)
library(cowplot)
# cbbPalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", 
#                 "#0072B2", "#D55E00", "#CC79A7", "#c5679b", "#be548f")

cbbPalette <- paletteer::paletteer_d(yarrr, info)
scales::show_col(cbbPalette)
```

# Plot of simulation result, one time run

Require run `1_npdr_cc.Rmd`, `1_npdr_qtrait.Rmd`.

## Get the statistics:

```{r}
alpha <- 0.05
n.variables <- 1000
pct.signals <- 0.1
true.pos <- test_df_qtrait %>%
  filter((pval.adj < alpha) & functional) %>%
  nrow()
n.pos <- test_df_qtrait %>%
  filter((pval.adj < alpha)) %>%
  nrow()
(npdr.recall.qtrait <- true.pos/(n.variables*pct.signals))
(npdr.precis.qtrait <- true.pos/n.pos)

alpha <- 0.05
true.pos <- test_df_cc %>%
  filter((pval.adj < alpha) & functional) %>%
  nrow()
n.pos <- test_df_cc %>%
  filter((pval.adj < alpha)) %>%
  nrow()
(npdr.recall.cc <- true.pos/(n.variables*pct.signals))
(npdr.precis.cc <- true.pos/n.pos)
```


## Generate figure 2:

Set theme:

```{r}
plot_r2 <- function(x){
  x + geom_smooth(method = "lm", se = F, color = 'grey') + 
  facet_wrap(~ method, scale = 'free_x', strip.position = 'bottom') +
  geom_point(aes(colour = functional), fill = NA, size = 2, shape = 21) +
  scale_x_continuous(breaks = scales::pretty_breaks(4)) + 
  theme_bw() + 
  theme(plot.title = element_text(hjust = 0.5, size = 12),
      strip.placement = 'outside',
      strip.switch.pad.wrap = unit(0, 'mm'),
      strip.text = element_text(size = 10.5),
      strip.background = element_rect(colour=NA, fill=NA),
      legend.position = c(0.66, 0.87),
      legend.key.width = unit(1, "line"),
      legend.key.height = unit(0.8, "line"),
      legend.text = element_text(size = 8),
      legend.title = element_text(size = 8),
      legend.margin = margin(c(1,1,1,1))) +
  scale_color_manual(values = cbbPalette[c(7,1)], labels = c('Non-functional', 'Functional')) +
  geom_text(cutoff_text, mapping = aes(x = x, y = y, label = label) , size = 3, color = cbbPalette[7]) + 
  geom_text(r2_text, mapping = aes(x = x, y = y, label = label), size = 3, parse = T, color = cbbPalette[7])
}
```

Now plot:

```{r}
# Continuous outcome:
qtrait_methods <-  c('RRelief score', 'Random forest score')
cc_methods <-  c('Relief-F score', 'Random forest score')
qtrait_r <- merge(test_df_qtrait, rf_npdr_qtrait) %>%
  dplyr::select(rrelief, beta.Z.att, functional, X.IncMSE) %>%
  rename(`RRelief score` = rrelief, `Random forest score` = X.IncMSE) %>%
  gather('method', 'imp_score', c(`RRelief score`, `Random forest score`))
qtrait_r$method <- factor(qtrait_r$method, levels = qtrait_methods)
cc_r <- merge(test_df_cc, rf_npdr_cc) %>%
  dplyr::select(rrelief, beta.Z.att, functional, MeanDecreaseAccuracy) %>%
  rename(`Relief-F score` = rrelief, `Random forest score` = MeanDecreaseAccuracy) %>%
  gather('method', 'imp_score', c(`Relief-F score`, `Random forest score`))
cc_r$method <- factor(cc_r$method, levels = cc_methods)


cutoff_text <- data.frame(
  label = rep('NPDR cutoff', 2), method = qtrait_methods,
  x = c(0.24, 18), y = c(npdr.beta.cutoff.qtrait + 4, npdr.beta.cutoff.qtrait + 4))
r2_text <- data.frame(
  label = c(deparse(bquote(r == .(r2_qtrait))), deparse(bquote(r == .(r2_rf_qtrait)))),
  method = qtrait_methods, x = c(0.25, 19), y = c(80, 80))
qtrait_plot <- ggplot(qtrait_r %>% arrange(functional), aes(x = imp_score, y = beta.Z.att)) + 
  labs(x = NULL, title = 'Continuous outcome data with main effect', 
       y = bquote('NPDR linear'~beta*minute), color = NULL) +
  geom_hline(yintercept = npdr.beta.cutoff.qtrait, linetype='dashed', color = cbbPalette[7])
qtrait_plot <- plot_r2(qtrait_plot)

# Dichotomous outcome:
cutoff_text <- data.frame(
  label = rep('NPDR cutoff', 2), method = cc_methods,
  x = c( 0.06, 3.7), y = c(npdr.beta.cutoff.cc + 1.25, npdr.beta.cutoff.cc + 1.25))
r2_text <- data.frame(
  label = c(deparse(bquote(r == .(r2_cc))), deparse(bquote(r == .(r2_rf_cc)))),
  method = cc_methods, x = c(0.062, 4) , y = c(24, 24))
cc_plot <- ggplot(cc_r %>% arrange(functional), aes(x = imp_score, y = beta.Z.att)) + 
  labs(x = NULL, title = 'Dichotomous outcome data with interaction effect', 
       y = bquote("NPDR logistic"~beta*minute), color = NULL) +
  geom_hline(yintercept=npdr.beta.cutoff.qtrait, linetype='dashed', color = cbbPalette[7])
cc_plot <- plot_r2(cc_plot)

fig1 <- plot_grid(qtrait_plot, cc_plot, nrow = 2, labels = 'AUTO')
fig1
ggsave(fig1, filename = 'figs/fig1.pdf', height = 5.5, width = 4.8)
```


## Generate figure 1:

```{r}
pr_qtrait <- prr_rf_qtrait %>%
  filter(type != 'NPDR') %>%
  rbind(prr_dat_qtrait) #%>% 

pr_cc <- prr_rf_cc %>% 
  filter(type != 'NPDR') %>%
  rbind(prr_dat_cc) #%>% 

pr_text <- data.frame(
  label = c(
    paste0('Random forest \n', c(round(auprc_rf_qtrait, 3), c(round(auprc_rf_cc, 3)))),
    paste0('RRelief \n ', round(auprc_relief_qtrait, 3)),
    paste0('Relief-F\n', round(auprc_relief_cc, 3)), 
    paste0('NPDR\n', round(auprc_npdr_qtrait, 3)),
    paste0('NPDR\n   ', round(auprc_npdr_cc, 3))), 
  sim = rep(c('Continuous outcome (main effect)', 'Dichotomous outcome (interaction)'), 3),
  x = c(0.3, 0.65, 0.485 , 0.28, 0.755, 0.81), 
  y = c(0.2, 0.2, 0.7, 0.91, 0.45, 0.88), 
  type = rep(c('Random forest', 'Relief-F', 'NPDR'), each = 2)
)

cutoff_point <- data.frame(
  sim = c('Continuous outcome (main effect)', 'Dichotomous outcome (interaction)'),
  x = c(npdr.recall.qtrait, npdr.recall.cc), y = c(npdr.precis.qtrait, npdr.precis.cc), 
  type = c('NPDR', 'NPDR')
)

pr <- rbind(pr_qtrait, pr_cc) %>% 
  mutate(sim = c(rep('Continuous outcome (main effect)', nrow(pr_qtrait)), 
                 rep('Dichotomous outcome (interaction)', nrow(pr_cc)))) %>%
  ggplot(aes(Recall, Precision, color = type)) +
  geom_path(size = 0.8) + theme_bw() + facet_wrap(~ sim) + 
  scale_color_manual(values = cbbPalette[c(1,4,2)]) +
  scale_x_continuous(limits = c(0, 1.05), breaks = seq(0, 1.1, 0.2), labels = scales::percent) +
  scale_y_continuous(limits = c(0, 1), breaks = seq(0, 1, 0.2), labels = scales::percent) +
  geom_text(pr_text, mapping = aes(x = x, y = y, label = label), size = 3) +
  geom_point(cutoff_point, mapping = aes(x = x, y = y), shape = 8, size = 2, stroke = 1) +
  coord_fixed(ratio = 1) +
  guides(color = FALSE)
pr
```


## Compare NPDR and STIR:
Require run `1_npdr_cc.Rmd`.
```{r}
p <- ggplot(npdr_stir, aes(x=stir.log10,y=npdr.log10)) + 
  geom_smooth(method = 'lm', se = F, color = 'grey') + 
  geom_point(aes(colour = functional), fill = NA, size = 2, shape = 21) +
  # geom_point(aes(colour = functional), alpha = 0.7, size = 3) +
  theme_bw() + theme(legend.position = c(0.78, 0.23)) +
  labs(colour = NULL, title = NULL,
       x = bquote('STIR ('~-log[10]~'p'~')'), 
       y = bquote('NPDR ('~-log[10]~'p'~')')) + 
  annotate("text", x = 130, y = 130, label = deparse(bquote(r > 0.99)), parse = T) +
  scale_color_manual(values = cbbPalette[c(7,1)], labels = c('Non-functional', 'Functional')) +
  geom_vline(xintercept=pcutoff, linetype='dashed', color = 'grey') +
  geom_hline(yintercept=pcutoff, linetype='dashed', color = 'grey')
p
ggsave(p, filename = paste0('figs/npdr_stir_p_cc.pdf'), height = 4.5, width = 4.5 )
```

# Replications of auPRC:
Require run `2_npdr_relief_100_cc.Rmd`, `2_npdr_relief_100_qtrait.Rmd`.

```{r}
load('results/npdr_100_cc.Rdata')
pr_cc_100 <- pr_df

load('results/npdr_100_qtrait.Rdata')
pr_qtrait_100 <- pr_df

cbbPalette <- paletteer::paletteer_d(yarrr, info)

# pr_df
pr_df <- rbind(pr_cc_100, pr_qtrait_100) %>%
  mutate(sim = c(rep('Dichotomous outcome (interaction)', nrow(pr_cc_100)), 
         rep('Continuous outcome (main effect)', nrow(pr_qtrait_100)))) %>%
  mutate(newtype = as.factor(case_when((type %in% c('Relief-F', 'RRelief')) ~ 'Relief',
                             TRUE ~ as.character(type))))
pr_df$newtype <- factor(pr_df$newtype, levels = c('Random forest', 'Relief', 'NPDR'))

pr_100 <- ggplot(pr_df, aes(x = newtype, y = aupr, color = newtype)) + 
  geom_boxplot(alpha = 0, color = 'grey50', outlier.size = -1, width = 0.5) +
  geom_jitter(height = 0, width = 0.15, alpha = 0.3) +
  labs(x = NULL, y = 'Area under the PR curve') + facet_wrap(~ sim) +
  theme_bw() + guides(fill = F, color = F) +
  scale_color_manual(values = cbbPalette[c(4,2,1)]) +
  scale_fill_manual(values = cbbPalette[c(4,2,1)]) +
  scale_y_continuous(
    breaks = seq(0, 1, 0.2), minor_breaks = seq(0.5, 1, 0.2),
    labels = scales::percent_format(accuracy = 1),
    limits = c(NA, 1.08)) + 
  ggsignif::geom_signif(
    comparisons = list(c('Random forest', 'NPDR'), c('Relief', 'NPDR')),
    map_signif_level = TRUE, y_position = c(1.06, 1.02), vjust = 0.5, tip_length = 0)
pr_100  
prs <- plot_grid(pr, pr_100, nrow = 2, labels = 'AUTO')
prs
ggsave(prs, filename = 'figs/fig2_pr_plots.pdf', width = 5.8, height = 6.4)

```

Dark theme for presentation:
```{r}
ggsave(pr_100 + ggdark::dark_theme_bw(), filename = 'figs/rocky_fig1.pdf', width = 5.8, height = 3.2)
ggsave(pr + ggdark::dark_theme_bw(), filename = 'figs/rocky_fig2.pdf', width = 5.8, height = 3.15)
```


```{r}
pr <- rbind(pr_qtrait, pr_cc) %>% 
  mutate(sim = c(rep('Continuous outcome (main effect)', nrow(pr_qtrait)), 
                 rep('Dichotomous outcome (interaction)', nrow(pr_cc)))) %>%
  filter(sim == 'Continuous outcome (main effect)') %>%
  ggplot(aes(Recall, Precision, color = type)) +
  geom_path(size = 0.8) + theme_bw() + 
  scale_color_manual(values = cbbPalette[c(1,4,2)]) +
  scale_x_continuous(limits = c(0, 1.05), breaks = seq(0, 1.1, 0.2), labels = scales::percent) +
  scale_y_continuous(limits = c(0, 1), breaks = seq(0, 1, 0.2), labels = scales::percent) +
  geom_text(pr_text %>% filter(sim == 'Continuous outcome (main effect)') , mapping = aes(x = x, y = y, label = label), size = 3) +
  geom_point(cutoff_point %>% filter(sim == 'Continuous outcome (main effect)'), 
             mapping = aes(x = x, y = y), shape = 8, size = 2, stroke = 1) +
  coord_fixed(ratio = 1) +
  guides(color = FALSE)
pr

pr_100 <- ggplot(pr_df %>% filter(sim == 'Continuous outcome (main effect)'), aes(x = newtype, y = aupr, color = newtype)) + 
  geom_boxplot(alpha = 0.5, color = 'grey50', outlier.size = -1, width = 0.5, fill = 'white') +
  geom_jitter(height = 0, width = 0.15, alpha = 0.3) +
  labs(x = 'Method', y = 'Area under the PR curve') +
  theme_bw() + guides(fill = F, color = F) +
  scale_color_manual(values = cbbPalette[c(4,2,1)]) +
  scale_fill_manual(values = cbbPalette[c(4,2,1)]) +
  scale_y_continuous(
    breaks = seq(0, 1, 0.2), minor_breaks = seq(0.5, 1, 0.2),
    labels = scales::percent_format(accuracy = 1),
    limits = c(NA, 1.08)) + 
  ggsignif::geom_signif(
    comparisons = list(c('Random forest', 'NPDR'), c('Relief', 'NPDR')),
    map_signif_level = TRUE, y_position = c(1.06, 1.02), vjust = 0.5, tip_length = 0)


pr_100
grant1 <- plot_grid(pr, pr_100, nrow = 1, labels = 'auto')
ggsave(grant1, filename = 'figs/grant_fig_1.pdf', width = 7, height = 3.2)

```


## auROC for Supplement:

```{r}
ss_df <- rbind(ss_df_cc, ss_df_qtrait) %>%
  mutate(sim = c(rep('Dichotomous outcome (interaction)', nrow(ss_df_cc)),
                 rep('Continuous outcome (main effect)', nrow(ss_df_qtrait))))

pr_text <- data.frame(
  label = c(
    paste0('Random forest \n', c(round(auroc_rf_cc, 3), round(auroc_rf_qtrait, 3))),
    paste0('Relief-F \n', round(auroc_relief_cc, 3)), 
    paste0('RRelief \n', round(auroc_relief_qtrait, 3)),
    paste0('NPDR ', round(auroc_npdr_cc, 3)), 
    paste0('NPDR\n', round(auroc_npdr_qtrait, 3))),
  sim = rep(c('Dichotomous outcome (interaction)', 'Continuous outcome (main effect)'), 3),
  x = c(0.6, 0.33, 0.8, 0.79, 0.83, 0.77), 
  y = c(0.63, 0.6, 0.38, 0.1, 1.05, 1), 
  type = rep(c('Random forest', 'Relief', 'NPDR'), each = 2)
)

roc <- ggplot(ss_df, aes(x = Sensitivity, y = Specificity, color = type)) +   
  geom_path(size = 0.8) + theme_bw() + facet_wrap(~ sim) + coord_fixed(ratio = 1) +
  scale_color_manual(values = cbbPalette[c(1,4,2)]) + guides(color = FALSE) +
  scale_x_reverse(breaks = seq(1, 0, -0.2), labels = scales::percent) + 
  scale_y_continuous(breaks = seq(0, 1, 0.2), labels = scales::percent, limits = c(NA, 1.05)) + 
  geom_text(pr_text, mapping = aes(x = x, y = y, label = label), size = 3)

roc
```


```{r}
recall_text <- data.frame(
  label = c('Random forest','Random forest',
            'RRelief', 'Relief-F',
            'NPDR', 'NPDR'),
  sim = rep(c('Continuous outcome (main effect)', 'Dichotomous outcome (interaction)'), 3),
  x = c(0.47, 0.53, 0.16, 0.17, 0.28, 0.05), 
  y = c(0.5, 0.8, 0.15, 0.15, 0.85, 0.99), 
  type = rep(c('Random forest', 'Relief', 'NPDR'), each = 2)
)

recall <- ggplot(ss_df, aes(x = pct_selected, y = Sensitivity, color = type)) +   
  geom_path(size = 0.8) + theme_bw() + facet_wrap(~ sim) + coord_fixed(ratio = 1) +
  scale_color_manual(values = cbbPalette[c(1,4,2)]) + guides(color = FALSE) +
  scale_x_continuous(breaks = seq(1, 0, -0.2), labels = scales::percent) + 
  scale_y_continuous(breaks = seq(0, 1, 0.2), labels = scales::percent) + 
  labs(x = 'Selection percentage cutoff', y = 'Recall') +
  geom_text(recall_text, mapping = aes(x = x, y = y, label = label), size = 3)

recall
ggsave(recall, filename = paste0('figs/recall_compare_1.pdf'), height = 3, width = 5)
```

# Replications of auROC for Supplement:

Must run `2_npdr_relief_100_cc_auroc.Rmd` and `2_npdr_relief_100_qtrait_auroc.Rmd`:

```{r}

load('results/npdr_100_cc_auroc.Rdata')
roc_cc_100 <- roc_df

load('results/npdr_100_qtrait_auroc.Rdata')
roc_qtrait_100 <- roc_df

cbbPalette <- paletteer::paletteer_d(yarrr, info)

roc_df <- rbind(roc_cc_100, roc_qtrait_100) %>%
  mutate(sim = c(rep('Dichotomous outcome (interaction)', nrow(roc_cc_100)), 
         rep('Continuous outcome (main effect)', nrow(roc_qtrait_100)))) %>%
  mutate(newtype = as.factor(case_when((type %in% c('Relief-F', 'RRelief')) ~ 'Relief',
                             TRUE ~ as.character(type))))
roc_df$newtype <- factor(roc_df$newtype, levels = c('Random forest', 'Relief', 'NPDR'))

roc_100 <- ggplot(roc_df, aes(x = newtype, y = aupr, color = newtype)) + 
  geom_boxplot(alpha = 0.5, color = 'grey50', outlier.size = -1, width = 0.5, fill = 'white') +
  geom_jitter(height = 0, width = 0.15, alpha = 0.3) +
  labs(x = NULL, y = 'Area under the ROC curve') + facet_wrap(~ sim) +
  scale_y_continuous(breaks = seq(0, 1, 0.1), 
                     minor_breaks = seq(0.65, 1, 0.1),
                     labels = scales::percent_format(accuracy = 1),
                     limits = c(NA, 1.05)) + 
  scale_color_manual(values = cbbPalette[c(4,2,1)]) +
  scale_fill_manual(values = cbbPalette[c(4,2,1)]) +
  ggsignif::geom_signif(
    comparisons = list(c('Random forest', 'NPDR'), c('Relief', 'NPDR')),
    map_signif_level = TRUE, y_position = c(1.045, 1.02), vjust = 0.5, tip_length = 0) +
  theme_bw() + guides(fill = F, color = F)
roc_100
rocs <- plot_grid(roc, roc_100, nrow = 2, labels = 'AUTO')
rocs

ggsave(rocs, filename = paste0('figs/roc_compare_100.pdf'), height = 6.24, width = 5)
```
