==> r/5a_visualize_mostafavi.Rmd <==
---
title: "Visualize results"
output: html_document
---

```{r}
library(tidyverse)
rm(list = ls())
set.seed(1618)
cbbPalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", 
                "#0072B2", "#D55E00", "#CC79A7", "#c5679b", "#be548f")

load('results/0_mostafavi_mdd_results.Rdata')
load('results/rf_0_mostafavi_mdd_results.Rdata')
```


```{r}
alpha_lab <- 0.001
npdr_w_wo_sex <- npdr_w_wo_sex %>%
  mutate(significant = as.factor((p.adj.npdr.nosex < alpha_lab) + (p.adj.npdr.sex < alpha_lab))) %>%
  mutate(sig.genes = ifelse((p.adj.npdr.nosex < alpha_lab) | p.adj.npdr.sex < alpha_lab, att, NA))

pboth <- ggplot(npdr_w_wo_sex, aes(x = nlog10.nosex, y = nlog10.sex)) + 
  geom_point(alpha = 0.7, shape = 21, size = 2.5, aes(fill = nlog10.sex.genes)) + 
  geom_hline(yintercept = -log10(0.05), linetype = 2, color = cbbPalette[1]) + 
  geom_vline(xintercept = -log10(0.05), linetype = 2, color = cbbPalette[1]) + 
  geom_abline(slope = 1, intercept = 0, alpha = 0.5, linetype = 3) +
  scale_x_continuous(breaks = seq(0, 14, 2)) +
  scale_y_continuous(breaks = seq(0, 14, 2)) +
  annotate("text", 13, -log10(0.05) + 0.2, label = "0.05 Bonferroni", 
           size = 3, color = cbbPalette[1]) +
  annotate("text", -log10(0.05) - 0.2, 9.7, label = "0.05 Bonferroni", 
           size = 3, color = cbbPalette[1], angle = 90) +
  labs(x = bquote('No-covariate NPDR ('~-log[10]~'p'[mdd]~')'),
       y = bquote('Sex-adjusted NPDR ('~-log[10]~'p'[mdd]~')'),
       fill = bquote(''~-log[10]~'p'['sex'])) +
  ggrepel::geom_text_repel(aes(label = sig.genes), #check_overlap = TRUE, 
            hjust=-0.12, vjust=1.7, fontface = "italic",
            seed = 1618, #angle = 45,
            size = 2, segment.alpha = 0.3) +
  viridis::scale_fill_viridis(option = 'B', direction = -1) + 
  theme_bw() + #theme(legend.position = c(0.9, 0.75)) +
  theme(legend.position = c(0.9, 0.8), 
      legend.key.height = unit(1, "line"),
      legend.key.width = unit(0.8, "line"),
      legend.text = element_text(size = 8),
      legend.title = element_text(size = 8)) +
  coord_fixed(ratio = 1) 
pboth
ggsave(pboth, filename = 'figs/mostafavi_npdrs_mdd.pdf', width = 6, height = 4.5)
```

Save figure with dark background for presentation:
```{r}
pboth_dark <- ggplot(npdr_w_wo_sex, aes(x = nlog10.nosex, y = nlog10.sex)) + 
  geom_point(alpha = 0.9, shape = 21, size = 3, aes(fill = nlog10.sex.genes)) + 
  geom_hline(yintercept = -log10(0.05), linetype = 2, color = cbbPalette[1]) + 
  geom_vline(xintercept = -log10(0.05), linetype = 2, color = cbbPalette[1]) + 
  geom_abline(slope = 1, intercept = 0, alpha = 0.5, linetype = 3) +
  scale_x_continuous(breaks = seq(0, 14, 2)) +
  scale_y_continuous(breaks = seq(0, 14, 2)) +
  annotate("text", 13, -log10(0.05) + 0.2, label = "0.05 Bonferroni", 
           size = 3, color = cbbPalette[1]) +
  annotate("text", -log10(0.05) - 0.2, 9.7, label = "0.05 Bonferroni", 
           size = 3, color = cbbPalette[1], angle = 90) +
  labs(x = bquote('No-covariate NPDR ('~-log[10]~'p'[mdd]~')'),
       y = bquote('Sex-adjusted NPDR ('~-log[10]~'p'[mdd]~')'),
       fill = bquote(''~-log[10]~'p'['sex'])) +
  ggrepel::geom_text_repel(aes(label = sig.genes), #check_overlap = TRUE, 
            # hjust=-0.12, vjust=1.7, 
            fontface = "italic",
            color = 'grey60',
            seed = 1618, #angle = 45,
            size = 2.5, segment.alpha = 0.3) +
  # viridis::scale_fill_viridis(option = 'B', direction = -1) + 
  # theme_bw() + #theme(legend.position = c(0.9, 0.75)) +
  theme(legend.position = c(0.9, 0.8), 
      legend.key.height = unit(1, "line"),
      legend.key.width = unit(0.8, "line"),
      legend.text = element_text(size = 8),
      legend.title = element_text(size = 8)) +
  coord_fixed(ratio = 1)  + 
  ggdark::dark_theme_bw() +
  viridis::scale_fill_viridis(option = 'D') 
ggsave(pboth_dark, filename = 'figs/mostafavi_npdrs_mdd_dark.pdf', width = 8, height = 6)
```


```{r}
npdr_rf_df <- npdr_w_wo_sex %>%
  dplyr::select(att, beta.npdr.sex, nlog10.sex.genes, sig.genes, p.adj.npdr.sex, p.adj.gene.sex) %>%
  merge(rf_imp_sorted, by = 'att') %>%
  arrange(p.adj.gene.sex) %>%
  mutate(hili = (p.adj.gene.sex < 1e-19)) %>%
  mutate(imp.genes = ifelse(hili, att, NA))

# rf_imp_sorted[rf_imp_sorted$att=='sex',]
npdr_rf_df_sorted <- npdr_rf_df %>% arrange(desc(beta.npdr.sex))
last.imp.att <- which(npdr_rf_df_sorted$p.adj.npdr.sex > 0.05)[1]-1
npdr.beta.cutoff.cc <- npdr_rf_df_sorted[last.imp.att, 'beta.npdr.sex']

npdr_rf_df %>% filter(MeanDecreaseGini > 0.1) %>% tally()

npdr_rf_df_nohili <- npdr_rf_df %>% filter(!hili)
npdr_rf_df_hili <- npdr_rf_df %>% filter(hili)
# x = importance
p_npdr_rf <- ggplot(npdr_rf_df_hili, aes(x = MeanDecreaseAccuracy, y = beta.npdr.sex)) + 
  geom_hline(yintercept = npdr.beta.cutoff.cc, linetype = 2, color = cbbPalette[1]) + 
  annotate("text", 2.7, npdr.beta.cutoff.cc + 0.4, label = "NPDR cutoff", 
           size = 4, color = cbbPalette[1]) +
  geom_point(npdr_rf_df_nohili, mapping = aes(x = MeanDecreaseAccuracy, y = beta.npdr.sex),
             shape = 21, size = 2.5, fill = alpha("grey", 0.5), color = alpha("grey", 0.3)) +
  geom_point(shape = 21, size = 2.5,
             mapping = aes(fill = nlog10.sex.genes)) +
  scale_x_continuous(limits = c(-2.5, 3)) +
  # scale_x_continuous(breaks = seq(0,0.5,0.02), limits = c(0, 0.26)) +
  scale_y_continuous(breaks = seq(-8,8,2)) +
  labs(x = bquote('Random forest permutation importance'),
       y = bquote('Sex-adjusted NPDR'~beta),
       fill = bquote(''~-log[10]~'p'['sex'])) +
  geom_text(aes(label = imp.genes), check_overlap = TRUE, 
            hjust=-0.15, vjust=1.15, fontface = "italic",
            size = 2.5) +
  viridis::scale_fill_viridis(option = 'B', direction = -1) + 
  theme_bw() + 
  theme(legend.position = c(0.92, 0.25), 
        legend.key.width = unit(0.8, "line"),
        legend.key.height = unit(1, "line"),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 8)) 

p_npdr_rf
# ggsave(p_npdr_rf, filename = 'figs/mostafavi_npdr_rf_mdd.pdf', width = 6, height = 6)
```

Random forest importance vs. nlog10.sex.genes:

```{r}
(r2.npdr.sex <- cor(npdr_rf_df$beta.npdr.sex, npdr_rf_df$nlog10.sex.genes))
(r2.rf.sex <- cor(npdr_rf_df$MeanDecreaseAccuracy, npdr_rf_df$nlog10.sex.genes))
npdr_spec <- data.frame(method = 'beta.npdr.sex', y = npdr.beta.cutoff.cc, label = 'NPDR cutoff')
npdr_rf_df %>%
  gather('method', 'score', c(MeanDecreaseAccuracy, beta.npdr.sex)) %>%
  ggplot(aes(x = nlog10.sex.genes, y = score)) + 
  facet_wrap(~ method, scales = 'free_y') +
  geom_hline(npdr_spec, color = cbbPalette[1],
             mapping = aes(yintercept = y), linetype = 2) + 
  geom_text(npdr_spec, color = cbbPalette[1],
            mapping = aes(x = 40, y = y + 0.4, label = label)) +
  geom_point(shape = 21, size = 2.5, fill = alpha("grey", 0.5), color = alpha("grey", 0.3)) + 
  labs(x = bquote(''~-log[10]~'p'['sex']),
       y = bquote('Importance score'),
       fill = bquote(''~-log[10]~'p'['sex'])) +
  viridis::scale_fill_viridis(option = 'B', direction = -1) + 
  theme_bw() + 
  theme(legend.position = c(0.92, 0.25), 
        legend.key.width = unit(0.8, "line"),
        legend.key.height = unit(1, "line"),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 8)) 
```







Summarize results:

```{r}
mydat <- npdr_w_wo_sex %>% 
  select(-c(p.adj.gene.sex, pval.gene.sex, beta.gene.sex)) %>%
  merge(gene_sex, by = 'att') %>%
  mutate(nlog10.sex.genes = -log10(p.adj.gene.sex))

best_genes <- mydat %>%
  filter(p.adj.npdr.sex < 0.05) %>%
  arrange(p.adj.npdr.sex) %>%
  select(att, p.adj.gene.sex, p.adj.npdr.sex, beta.npdr.sex, nlog10.sex.genes)

best_genes %>% tally()
best_genes %>% filter(p.adj.gene.sex < 0.05) %>% tally()

best_genes_nosex <- npdr_w_wo_sex %>%
  filter(p.adj.npdr.nosex < 0.05) %>%
  arrange(p.adj.npdr.nosex) %>%
  select(att, p.adj.npdr.nosex, pval.npdr.nosex, 
         beta.npdr.nosex, beta.raw.att.x,
         pval.gene.sex, p.adj.gene.sex)

best_genes_nosex %>% tally()
best_genes_nosex %>% filter(p.adj.gene.sex < 0.05) %>% tally()

sex_genes <- npdr_w_wo_sex %>%
  mutate(p.bon.gene.sex = p.adjust(pval.gene.sex, method = 'bonferroni')) %>%
  select(att, beta.gene.sex, pval.gene.sex, p.bon.gene.sex) %>%
  filter(p.bon.gene.sex < 0.05) %>%
  arrange(p.bon.gene.sex)

write_csv(sex_genes, 'results/mostafavi_sex_genes.csv')
write_csv(best_genes, 'results/mostafavi_best_0.05_npdr_genes.csv')
write_csv(best_genes_nosex, 'results/mostafavi_best_0.05_npdr_genes_nosex.csv')

```
