==> r/5b_visualize_jerzy.Rmd <==
---
title: "Visualize results"
output: html_document
---

```{r}
library(tidyverse)
# rm(list = ls())
set.seed(1618)
cbbPalette <- c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", 
                "#0072B2", "#D55E00", "#CC79A7", "#c5679b", "#be548f")

load('results/0_jerzy_mdd_results.Rdata')
load('results/rf_0_jerzy_mdd_results.Rdata')
```


```{r}
r2_p_jerzy <- round(cor(npdr_w_wo_sex$nlog10.nosex, npdr_w_wo_sex$nlog10.sex), 3)
pboth <- ggplot(npdr_w_wo_sex, aes(x = nlog10.nosex, y = nlog10.sex)) + 
  geom_point(alpha = 0.7, shape = 21, size = 2, aes(fill = nlog10.sex.genes)) + 
  geom_hline(yintercept = -log10(0.01), linetype = 2, color = cbbPalette[1]) + 
  geom_vline(xintercept = -log10(0.01), linetype = 2, color = cbbPalette[1]) + 
  geom_abline(slope = 1, intercept = 0, alpha = 0.5, linetype = 3) +
  annotate("text", 9.8, -log10(0.01) + 0.3, label = "0.01 Bonferroni", 
           size = 3, color = cbbPalette[1]) +
  annotate("text", -log10(0.01) - 0.3, 9.8, label = "0.01 Bonferroni", 
           size = 3, color = cbbPalette[1], angle = 90) +
  labs(x = bquote('No-covariate NPDR ('~-log[10]~'p'[mdd]~')'),
       y = bquote('Sex-adjusted NPDR ('~-log[10]~'p'[mdd]~')'),
       fill = bquote(''~-log[10]~'p'['sex'])) +
  geom_text(aes(label = sig.genes), check_overlap = TRUE, 
            hjust=1.15, vjust=0, fontface = "italic",
            #seed = 1618, #angle = 45, segment.alpha = 0.3
            # nudge_x = 0.5, nudge_y = 0.5,
            size = 2) +
  viridis::scale_fill_viridis(option = 'B', direction = -1) + 
  annotate("text", x =  8.5, y = 11.5, label = deparse(bquote(R^2 > 0.999)), parse = T) +
  theme_bw() + #theme(legend.position = c(0.9, 0.75)) +
  theme(legend.position = c(0.9, 0.65), 
      legend.key.height = unit(0.7, "line"),
      legend.key.width = unit(0.4, "line"),
      legend.text = element_text(size = 6),
      legend.title = element_text(size = 6)) +
  coord_fixed(ratio = 1) 
pboth
ggsave(pboth, filename = 'figs/jerzy_npdrs_mdd.pdf', width = 4, height = 3.8 )
```


```{r}
npdr_rf_df <- npdr_w_wo_sex %>%
  select(att, beta.npdr.sex, nlog10.sex.genes, sig.genes, p.adj.npdr.sex) %>%
  merge(rf_imp_sorted, by = 'att') %>%
  mutate(imp.genes = ifelse((p.adj.npdr.sex < 0.01) | 
                            (MeanDecreaseGini > 0.2) |
                            (p.adj.npdr.sex < 0.05), att, NA)) 

p_npdr_rf <- ggplot(npdr_rf_df, aes(x = MeanDecreaseGini, y = beta.npdr.sex)) + 
  geom_point(alpha = 0.7, shape = 21, size = 1.5, aes(fill = nlog10.sex.genes)) + 
  geom_hline(yintercept = 0, linetype = 2, color = cbbPalette[1]) + 
  scale_x_continuous(breaks = seq(-1,1,0.2)) +
  scale_y_continuous(breaks = seq(-8,8,2)) +
  labs(x = bquote('Random forest Gini importance'),
       y = bquote('Sex-adjusted NPDR'~beta),
       fill = bquote(''~-log[10]~'p'['sex'])) +
  geom_text(aes(label = imp.genes), check_overlap = TRUE, 
            hjust=-0.12, vjust=0.1, fontface = "italic",
            size = 2.5, segment.alpha = 0.3) +
  viridis::scale_fill_viridis(option = 'B', direction = -1) + 
  theme_bw() + 
  theme(legend.position = c(0.8, 0.08), 
        legend.key.width = unit(1, "line"),
        legend.key.height = unit(0.8, "line"),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 8),
        legend.direction = 'horizontal')
p_npdr_rf
ggsave(p_npdr_rf, filename = 'figs/jerzy_npdr_rf_mdd.pdf', width = 6, height = 4)

```


Summarize results:

```{r}
best_genes <- npdr_rf_df %>%
  filter(p.adj.npdr.sex < 0.05) %>%
  arrange(p.adj.npdr.sex) %>%
  select(-c(HC, MDD))

sex_genes <- npdr_w_wo_sex %>%
  mutate(p.bon.gene.sex = p.adjust(pval.gene.sex, method = 'bonferroni')) %>%
  select(att, beta.gene.sex, pval.gene.sex, p.bon.gene.sex) %>%
  filter(p.bon.gene.sex < 0.05) %>%
  arrange(p.bon.gene.sex)

write_csv(sex_genes, 'results/jerzy_sex_genes.csv')
write_csv(best_genes, 'results/jerzy_best_0.05_npdr_genes.csv')
```
