==> r/6_tab_psqi.Rmd <==
---
title: "Summarize PSQI results"
output: html_document
---

```{r}
gene_stat <- read_csv('results/jerzy_psqi_best_0.05_npdr_genes.csv')
gene_desc <- read_csv('results/psqi_genes_desc.csv')
gene_sum <- merge(gene_stat, gene_desc, by = 'att') %>%
  mutate(description = tolower(Description)) %>%
  select(att, description, beta.npdr.age, pval.npdr.age, p.adj.npdr.age) %>%
  arrange(p.adj.npdr.age)
write_csv(gene_sum, 'results/psqi_gene_tab.csv')
```
